[mojo-stdlib] Localize handling of `__disable_del` with helper methods.
The `__disable_del` utility is a very sharp object that disables the
destructor of a type.  We don't really want code to randomly disable
the destructor of some type that they don't control, because this
exposes coupling of the behavior of that type with the caller.

This patch improves the standard library by introducing a couple of
methods (e.g. `VariadicPack.consume_elements`) to encapsulate this,
providing a much safer and easier to use API for the purposes used
before.

The consequence of this is that `__disable_del` is now only used on
`self` in a consuming method, which is much safer.

MODULAR_ORIG_COMMIT_REV_ID: 978c76efe17b8c00b1eaaa0bf526ad7ab5523777