[stdlib] feat: Use `ExplicitlyCopyable` for List, Span, and InlineArray element types
This narrows the trait bound for several collection element types from
`Copyable` to merely `ExplicitlyCopyable`.

This is possible with a relatively small diff because `Copyable`
requires `ExplicitlyCopyable` (as of the previous PR), meaning that many
bounds out in the wild using `Copyable` can still be used with
collections that require only `ExplicitlyCopyable`.

The following parameters were updated to require only
`ExplicitlyCopyable`:

* `List.T`
* `_ListIter.T`
* `Span.T`
* `_SpanIter.T`
* `InlineArray.ElementType`
* `UnsafePointer.init_pointee_copy()`

This exposed a handful of places where implicit copies needed to be
updated to explicit copies in the implementations of `List` and `Span`.

MODULAR_ORIG_COMMIT_REV_ID: 3d758c3cdb95a4a2160b9b44e9a6afa8c4528731