[Scheduler] Refactor batch construction logic into separate class
This should be NFC. I moved the logic that constructs an optimal batch
out of text_generation_scheduler.py into a separate file. We intend to
reuse this TextBatchConstructor for the Decode and Prefill schedulers.
This will allow for greater code reuse and less duplication.

Even though this change is big and scary, I have good confidence in
correctness since the `test_paged_scheduler` tests are pretty thorough.

Note that the APIs differ a bit from proposed in the Scheduler 2.0 doc.
This is intentional. I make the simplest API possible in this PR that
minimizes drift from existing code. Future PRs can make things nicer and
close to the desired state.

Note that I uncovered a bug in this exercise. We were scheduling CE reqs
in the incorrect order. Instead of scheduling the earliest arriving /
oldest
request first, we were scheduling the latest arriving / newest request
first.
This is a bug that has been here for a while. This doesn't really matter
for infinite request rate benchmarks, but may have real impact on TTFT
in
other scenarios.

Fixing this involves toggling a boolean (popitem(last=True) vs
popitem(last=False)).
However, we will have to update a bunch of tests so that is deferred a
bit.

My priority ATM is to connect TextBatchConstructor to the Decode and
Prefill
schedulers and get Llama 405B churning along smoothly.

MODULAR_ORIG_COMMIT_REV_ID: b39e6cc24bba16299af6737ce64746c48a235cae