[docs] Remove Why Mojo page, merge info into FAQ
This page was useful in the early days to explain in detail why we
created Mojo, but after 2+ years, it’s not necessary to elaborate on a
dedicated page. A lot of the reasons were already included in the FAQ,
so this simply merges and expands some of those answers with details
from the Why Mojo page, and discards the rest that’s not necessary
anymore.

The Why Mojo page now redirects to the "Motivation" section of the FAQ.

MODULAR_ORIG_COMMIT_REV_ID: 5b4be66ea3075254a6062a3c891073fcc4f81df8