[SDK] Require DType input to ops.range
By explicitly setting a DType it's much easier to infer user intent for
the output of `range`. The Graph API tries to be simple, predictable,
and explicit, and explicit DTypes aid greatly in this.

As a result the logic for ranges was simplified greatly, and many cases
which previously required explicitly creating a constant or specifying
an output dimension no longer do.

MODULAR_ORIG_COMMIT_REV_ID: 0071dd7142b1d450f82273ddedeb7cc45600483c