[SDK] Bringup GPT OSS
This is the main pr that brings up most of the model. It has the
following layers:

* RMS norm: multiply before cast like Gemma3
* Yarn Rotary Position Embeddings: It’s not the same as Llama at all
because the model config specifies rope type to be Yarn for GPT-OSS and
llama3 for Llama3. The math is very different.
* TransformerBlock: we can use a generic distributed transformer block
that accepts an MLP of type nn.Module (it will be GPT_MOE) and an
attention layer of type nn.Module which will be GPT_OSS_Attention
* GPT_OSS_Attention layer has sink parameters and generates the mask
depending on the layer type (full attention or local window attention
supported).

MODULAR_ORIG_COMMIT_REV_ID: 98b62310cebe172224e39301ebed1338c806746d