[stdlib] Remove `run_destructors` from `InlineArray`
Now that we have a way to ask a type if it is trivially destructible,
remove the keyword-parameter from `InlineArray` that allowed callers to
explicitly control destructors. Instead, query the element type's
property of whether it is trivially destructible. If it is not, run the
destructor on each of the elements in the `InlineArray`. Otherwise, it's
a no-op.

MODULAR_ORIG_COMMIT_REV_ID: 18448ee68a11a50d6eab23324c9ea4accc3e87a5