[Serve] Fix mutability bug in metrics emission
Fixes a bug relating to our detailed metrics emission. Before this, we
could clear/mutate our buffer of detailed metrics before the queue
thread gets the chance to serialize and send the list. Additionally, all
detailed metrics lists we send to the queue would be pointing to the
same reference, resulting in dropped metrics and double counted metrics.

This code change just ensures that we create a new list before sending
it to the queue, so we'll never mutate the reference after sending the
object to the queue.

A few other tangential bugs that needed to be fixed to enable tests:
- Added a knob to set the buffering factor of detailed metrics to ensure
  we immediately get detailed metrics during testing
- Explicitly propagate settings objects from calling process to
  subprocess when initializing our metrics client
- Remove early returns from our log_metrics method to ensure we still
  send metrics even if we don't want to log.

MODULAR_ORIG_COMMIT_REV_ID: 1324daabffc97c1473b49fe10a374a79e0f92d69