[mojo-lang] Fix Dict getitem emission mutability problem.
When emitting a dynamic lvalue argument in a reference-requiring
operand to a call, we generally bind values into a temporary and
bind an immutable reference to that temporary.  That makes sense
except in the case where we have a dynamic lvalue with a getitem that
returns a mutable reference.  In this case, we don't want to throw
away mutability of the argument.

This Fixes https://github.com/modular/modular/issues/4705 and
Fixes MOCO-2077, which should improve ergonomics around Dict.

MODULAR_ORIG_COMMIT_REV_ID: b860aa56fbe469811c65c0a88620d75148b0174b