[stdlib] Input and output types of multimem ops must be equal
The output type was being set to to the accumulation type in the
multimem funtions. When the input was bfloat16 and the accumulation type
was float32 this was giving wrong results. The output type must be the
same as the input type. The accumulation type is only an intermediate
results for the reduction. After the reduction the output is cast back
to the input type. This is done by the instruction.

MODULAR_ORIG_COMMIT_REV_ID: 00ce9d94230862d55b394b3ebdd9fdefc5a82aad