[External] [stdlib] Fix `Atomic.compare_exchange` and add tests. (#68032)
[External] [stdlib] Fix `Atomic.compare_exchange` and add tests.

This fixes some current bugs and UB in the previously named
`Atomic.compare_exchange_weak`.

1. Updates name to `compare_exchange`: The previous name indicated that
this function was the `weak` variant of compare exchange which is was
*not*. The compiled llvm instructions show the use of `cmpxchg` without
the `weak`. We should add the weak variant in the future, but first we
need to have the `pop.atomic.cmpxchg` MLIR operation support the `weak`
attribute.
2. Correctly update the `expected` argument for floating-point atomics:
Before, if an `Atomic`'s `DType` was a floating point, the inout
parameter `expected` would not be correctly updated if the `cmpxchg`
failed.
3. Fix UB when writing the atomic's loaded value into the `expected`
argument: Previously when the `cmpxchg` operation was called, if the
success flag was `False`, the expected inout parameter would be written
to by non-atomically reading the atomic pointer via `expected =
value_addr[]`. This is UB as we are not allowed to read/write to the
atomic pointer without an atomic operation. Instead, now we extract the
loaded value from the result of the `cmpxchg` operation and store that
in `expected`.

ORIGINAL_AUTHOR=Nathan Ward <43621845+NathanSWard@users.noreply.github.com>
ORIGINAL_USER=@NathanSWard

Co-authored-by: Nathan Ward <43621845+NathanSWard@users.noreply.github.com>
Closes modular/modular#5223
MODULAR_ORIG_COMMIT_REV_ID: cbf3799ae2e74ee677fe8c1e179848b3401f2a0a