[SDK] Elementwise mul with fastmath=none
Elementwise multiply with fastmath=none to fix a numerical issue.
The specific issue is when you have y + x*y when y ~= 1, x ~= -1, and
dtype bfloat16.
This leads to a difference in units in the last place of up to 127,
an error that downstream ops such as layer norm can amplify.
So remove the fastmath contraction flag to prevent FMA, bringing our
numerics in line with PyTorch (which models are trained in).

Testing:
- Added a test that max.graph `ops.add(ops.mul(...), ...)` is within one
  unit in the last place of torch `a * b + c` on GPU.
MODULAR_ORIG_COMMIT_REV_ID: a85a52981eeafcd75d7a6e732c5bd85ac6bcc4d8