[Kernels] Provide more information when cublas libraries are not found
Not all of our container images contain cublas, but some benchmarks use
cublas for competitive analysis. This provides more informative error if
the library is not found.

ERROR: Cannot find the cuBLAS libraries. Please make sure that the CUDA
toolkit is installed and that the library path is correctly set in one
of the following paths [libcublas.so.14]. You may need to make sure that
you are using the non-slim version of the MAX container.
MODULAR_ORIG_COMMIT_REV_ID: 3e52101de409da30bdee12149662f9bf0e410eaa