[Stdlib][Kernels] Do not crash if dylib not found
We sometimes use external libraries and currently the default behavior
is to crash if the library cannot be found (via abort). This is ok when
running in a mojo session, but not ok when running in a max session
since we do not want to crash the server. This plumbs things so that we
raise when a dylib is not found and that exception can then be caught be
the max inference session.

MODULAR_ORIG_COMMIT_REV_ID: 9c53350dea3fda1aad534455aee71e42ee39bdd6