[Interfaces] Move EmbeddingsGenerator to Pipeline Interface
In a similar fashion to the retirement of the TokenGenerator interface,
we have moved the EmbeddingsGenerator to the general Pipeline interface.
This is intended to provided the opportunity for simplification to MAX
Serve, at the Scheduler & API layer.

MODULAR_ORIG_COMMIT_REV_ID: d127b5672bda099b7cba4e7b707a6d2fefa9b031