[SDK] Only move logit offsets to CPU for non-batch logit processors
The `batch_logit_offsets` is used in two ways:
1) To index into the logits to pass individual logits into each logits
processor. For this, the offsets must be on CPU.
2) `batch_logit_offsets` is passed to the batch processors as-is (on GPU
if available).

There was a bug since we were forcing the batch logit offsets to be
moved to CPU when it should still be on GPU for the second case.

MODULAR_ORIG_COMMIT_REV_ID: 3147a57957ee1143ce515cea7e190bb4d9d23bad