[Pipelines] Error early if no room for kv cache
This handles the case where the model and activations are _just_ big
enough to take all of the device memory without leaving room for the
KVCache. We would technically catch this when constructing the paged
cache, but it's better to catch it here.

MODULAR_ORIG_COMMIT_REV_ID: 125aef108f69df77c5e287d64f9233704e4435a1